Needle Craft
Needlecraft encompasses a variety of crafts that involve using a needle to manipulate thread, yarn, or other materials to create decorative or functional items. It includes techniques like embroidery, knitting, crochet, needlepoint, and tatting. Essentially, needlecraft is the art of working with a needle to create textiles and other decorative or functional pieces.